Deux petits tests - Reglages 2 pb : Menu centre sur tel / Bug sans saisie de date booking
