+++ app/assets/stylesheets/components/_navbar.scss
@@ -5,7 +5,7 @@
   justify-content: space-between;
   padding: 0.5rem 1rem;
   background-color: $black;
-  .navbar-brand img {
+  .navbar-brand img, .width-70-px {
     width: 70px;
   }
   .loggin {
