+++ app/views/shared/_navbar.html.erb
@@ -58,7 +58,7 @@
       <% if current_user.bookings.select { |booking| booking.state == "pending" }.present? %>
         <%= link_to t('navbar.pay-btn'), allmybookings_path, class: "button-access-book" %>
       <% end %>
-      <div class="nav-item dropdown">
+      <div class="nav-item dropdown width-70-px">
 
           <%= image_tag "log-in.png", class: "avatar dropdown-toggle", id: "navbarDropdown", data: { toggle: "dropdown" }, 'aria-haspopup': true, 'aria-expanded': false %>
 
@@ -75,7 +75,7 @@
     <%end%>
   </div>
 
-  <div class="collapse navbar-collapse" id="navbarSupportedContent" style="<%= 'margin-right: 65px' if current_user %>">
+  <div class="collapse navbar-collapse" id="navbarSupportedContent" style="<%= 'margin-right: 65px' if (user_signed_in? && current_user.bookings?) %>">
     <ul class="navbar-nav mr-auto">
       <% if user_signed_in? %>
         <li class="nav-item <%= 'active-pattaya' if params[:controller] == 'pages' && params[:action] == 'home'%>">
